DHTML Menus initial version.
The DHTML Menus module add Javascript to the hierarchical menus to remove
some of the page refreshes. They are not hover menus like many DHTML
setups, but instead they click to expand and then click to close again,
as is expected from a typical tree browser. Any menu block can be activated
with DHTML abilities.